Added id field to Timer class for differentiating timers.

+++ Board/src/Board.hpp
@@ -53,7 +53,7 @@ public:
    virtual void deepSleep(
       const long& milliseconds) = 0;
 
-   virtual int systemTime() const = 0;
+   virtual long systemTime() const = 0;
 
 private:
