+++ Board/src/Esp8266Board.cpp
@@ -65,7 +65,7 @@ void Esp8266Board::deepSleep(
    // TODO
 }
 
-int Esp8266Board::systemTime() const
+long Esp8266Board::systemTime() const
 {
    return (millis());
 }
